refactor(step3): 테스트 코드 중 Car과 Game 클래스를 분리
